# api_gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for shorten urls",
        "title": "CoolURLShortener API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/save_url": {
            "post": {
                "description": "Принимает исходную ссылку, создает короткую ссылку и возвращает короткую ссылку",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url"
                ],
                "summary": "Создание и сохранение короткой ссылки по исходной ссылки",
                "operationId": "save-url",
                "parameters": [
                    {
                        "description": "Длинная ссылка",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LongURLData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.URlData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    }
                }
            },
            "options": {
                "description": "Возвращает информацию по хедерам Access-Control-Request-Method, Access-Control-Request-Headers, Origin",
                "tags": [
                    "options"
                ],
                "summary": "Получение описания параметров соединения с сервером",
                "operationId": "options-save-url",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/top_urls": {
            "get": {
                "description": "Принимает page и limit. Возвращает список популярных url. Поддерживает пагинацию",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url"
                ],
                "summary": "Получение списка популярных url",
                "operationId": "get-top-urls",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Страница",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Максимальное количество url на странице",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TopURLDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    }
                }
            }
        },
        "/{short_url}": {
            "get": {
                "description": "Принимает короткую ссылку в path параметрах и производит редирект на исходную ссылку",
                "tags": [
                    "url"
                ],
                "summary": "Редирект с короткой ссылки на исходную ссылку",
                "operationId": "follow-url",
                "parameters": [
                    {
                        "type": "string",
                        "description": "короткая ссылка",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Body"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.LongURLData": {
            "type": "object",
            "required": [
                "long_url"
            ],
            "properties": {
                "long_url": {
                    "type": "string"
                }
            }
        },
        "dto.Pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "next": {
                    "type": "integer"
                },
                "previous": {
                    "type": "integer"
                },
                "record_per_page": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "dto.TopURLData": {
            "type": "object",
            "properties": {
                "create_count": {
                    "type": "integer"
                },
                "follow_count": {
                    "type": "integer"
                },
                "long_url": {
                    "type": "string"
                },
                "short_url": {
                    "type": "string"
                }
            }
        },
        "dto.TopURLDataResponse": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/dto.Pagination"
                },
                "top_url_data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TopURLData"
                    }
                }
            }
        },
        "dto.URlData": {
            "type": "object",
            "properties": {
                "long_url": {
                    "type": "string"
                },
                "short_url": {
                    "type": "string"
                }
            }
        },
        "response.Body": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}